Added a project to build the Client for Silverlight 5.
The existing Silverlight Client cannot be used with Silverlight 5 because it uses the System.Threading.Tasks NuGet Package which conflicts with the official implementation of Tasks in Silverlight 5.

The System.Threading.Tasks NuGet package does include a build for Silverlight 5 which uses the official Tasks api, but it also includes a TaskExtensions class containing the Unwrap method, which still conflicts with the official TaskExtensions class. So the System.Threading.Tasks NuGet can't be used with SignalR.Client at present.

Instead, this Check-in includes a stand-in (and very naive) implementation of ConcurrentDictionary which is the only part of the TPL used by SignalR.Client and missing from Silverlight 5.