Prevented caching ajax requests on reconnect.
This commit fixes an issue when SignalR triest to reconnect after a webserver has returned an error page with client side caching enabled.

In this case, SignalR caches the response of the webserver, which results in many requests per second being fired but each request results with a 304 (not modified status code), without ever accessing the server.

In case of IE7-IE9, this will eventually crash the browser.

It's solved by adding a 'cache:false' in the pollXhr ajax parameters (already present in the negotiate).