Added OnRejoiningGroups to PersistentConnection
Fixes Issue #525

OnRejoiningGroups takes a list of groups a reconnecting client claims to be a
member of and returns a list of groups the client should actually rejoin.

The default implementation of OnRejoiningGroups always returns an empty
list of groups since the list of groups a client claims to be a member of
cannot be trusted.

Hub.RejoiningGroups is analogous to PersistentConnection.OnRejoiningGroups