Added logic in reconnect to detect if the connection has been forcibly disconnected.
The reason why this logic is necessary is that the onclose for websocket
is what triggers a reconnect.  However if we forcibly stop a connection
via $.connection.hub.stop() the websockets reconnect gets triggered.
This is expected because we close the socket.  This is also what helps
websockets reconnect on network loss which is why we can't just add the
logic in the onclose handler.

#823