IHubactivator now creates HubActivationResults and decoupled deactivating from the dispatcher
- IHubActivator now creats HubActivationResults which holds a IHub and
an additional NVC Items which holds metadata about the creation of the
hub (which happens in the IHubActivator). This is useful for saving a
state (like a lifetimescope)
- Separated the deactivating (destruct) of the hub in a seperate class
so this is overridable. The reason to make this overridable is that the
user can clean up his items which are created in the IHubActivator.