Added browser sniffing to our unit tests.
- It's impossible to detect what version of web sockets the browser
supports without attempting to start a websocket connection, therefore
we know that the web sockets in ios devices that are not ios6 do not
support our protocol.
- ForeverFrame can currently attempt to run in android devices but
fails, therefore we need to disable it for devices like that.  We
created forever frame for the sole purpose of being used in IE so added
a check to verify its correctness

#1359