HostIlE - IE11 connects to a wrong host in cross domain scenarios when a relative connection url is used (fixes #2959)
When parsing a relatinve uri with <a /> (anchor) IE11 does not provide the protocol so we fall back to using `window.document.location` to resolve connection details. However if the host could be read from the url provided by the user we should use it instead of the one from the `window.document.location`. The fix is to use the host from the url provided by the user if possible and fall back to using the host from the `window.document.location` if the host could not be read from the user's url.