BetterThreadSafeThanSorry - fixing UrlBuilder threading issue
The same instance of UrlBuilder could be used by multiple threads which could corrupt the internal state of the StringBuilder instance used by the UrlBuilder which results in weird exceptions. This manifested as random CI build failures. The fix is to turn the UrlBuilder to a static class (it does not need to be mocked) and create a new instance of StringBuilder each time.