Transport!YouAreFinished! - generalizing AbortSent flag
The transport is not in a usable state if either the abort request is sent or the transport was disposed. When in this state the transport should no longer be used. Before it was guarded by the AbortSent flag but the abort request does not necessarily have to be sent so the AbortSent was replaced with a new Finished flag which is set to true if either the Abort method was called or the transport was disposed.