YouFailedToFail.JS - complete "connect" request if websocket closed
If the server closed the websocket while the client was waiting for a response to the "connect" request the client would not notice this and would wait until the connect request timed out. The fix is to explicitly finish the initialization phase if it is in progress and the websocket is closed.

Fixes #2902