SleepWell - adding a workaround for a race condition in WebSockets
There seems to be a race condition in WebSockets which may result in a deadlock/livelock which in turn causes a failure to the EndToEndTests. Adding a delay seems to fix the issue so we can use it as a workaround until the issue in the WebSocket implementation is fixed.