WhoFliesHighFallsLow - pushing TransportInitializeHandler to ClientTransportBase
Removing references to TransportInitializationHandler from transport implementations since initialization is now handled in the ClientTransportBase class. Types derived from the `ClientTransportBase` can now override the OnStartFailed protected method to be notified if/when a failure occurs during starting the transport and can invoke the `TransportFailed` method if a transport specific error occurred to notify the `ClientTransportBase` about the error.