HopeIsTheOnlyStrategy - fixing hangs when creating Perf Counters
The theory is that tha hangs are caused by a CultureInfo that leaked
from a different AppDomain (see this blog post for more details:
(http://www.zpqrtbnk.net/posts/appdomains-threads-cultureinfos-and-paracetamol).
Unfortunately, I was not however able to reproduce the issue (tried both English
and non-English Windows) but hope it will fix #3414 as indicated in the discussion
and the above blog post.