fix #4152 by adding '/configFile' to ghp command
fixes #4152

This change adds support for a new '/configFile:' option to the 'signalr.exe' tool's 'ghp' command

If your server app uses a later version of an assembly than the 'signalr.exe' tool uses, we need a binding redirect. Newtonsoft.Json is the most common trigger here.

To fix this, you need to apply the binding redirect to the AppDomain that 'signalr.exe' creates to invoke the JavaScriptGenerator. This means you either need to put it in a 'signalr.exe.config' next to 'signalr.exe' (yuck), or we need to provide a different mechanism. This change makes it possible to use the '/configFile:' option to pass in the path to your app's Web.config file, which should already contain the necessary redirects.