fix #4207 by adding 'Error' to negotiate response (#4245)
fixes #4207

This adds an "Error" field to the negotiate payload that contains an
error message provided by the server. If present, the `.start`/`.Start`
Task/Promise will be faulted with that error. This only works with
clients that support Protocol version 2.0, so it also has to have the
ProtocolVersion 2.0 value. This is designed to prevent older clients
from continuing to attempt to connect when they get the error response.
Since we use a new ProtocolVersion, older clients will fail with the
protocol incompatibility error.