fix #4213 by tracking callback identity better (#4249)
fixes #4213

Previously we used the callback as the key in a JavaScript object. ECMAScript objects are required to have 'String' keys, which means the function is stringified and the string representation is used as the key. The problem is that functions can often have the same string representation. For example, if they are wrapped in something like `bind` or `apply`, the representation is `function() { [native code] }`.

This change switches to use an array and a 'guid' assigned by 'on' as the lookup system. That avoids the problems with stringifying functions. This system of assigning a 'guid' is used by jQuery internally for it's own event lists so it seems like the most appropriate way to handle this.