fix #4152 by adding '/configFile' to ghp command (#4247)
fixes #4152

This change adds support for a new '/configFile:' option to the 'signalr.exe' tool's 'ghp' command

If your server app uses a later version of an assembly than the 'signalr.exe' tool uses, we need a binding redirect. Newtonsoft.Json is the most common trigger here.

To fix this, you need to apply the binding redirect to the AppDomain that 'signalr.exe' creates to invoke the JavaScriptGenerator. This means you either need to put it in a 'signalr.exe.config' next to 'signalr.exe' (yuck), or we need to provide a different mechanism. This change makes it possible to use the '/configFile:' option to pass in the path to your app's Web.config file, which should already contain the necessary redirects.

Our own build tests this change! I updated the Samples project to JSON.NET 11 and immediately hit this issue. When building the Samples project we use `signalr.exe ghp` to generate hub proxies. With these changes, if the `/configFile` option is **not** specified, the build fails with this error message:

```
5>EXEC : error : Failed to load assembly 'Newtonsoft.Json, Version=6.0.0.0,
Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed' referenced by the server. If your
application uses a newer version of this assembly, this error is likely due to a missing
or inaccurate binding redirect. If your server application has a 'Web.config' or
'[AppName].exe.config' file with binding redirects in it, provide the path to that file
with the '/configFile' parameter. If you have already provided a '/configFile' parameter,
make sure the config file it points to has a binding redirect for 'Newtonsoft.Json,
Version=6.0.0.0, Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed' within it. See
https://aka.ms/about-binding-redirects for more information on binding redirects.
```

I added the above error message to help guide users when they encounter this issue because it is not an obvious fix :).

Once the `/configFile` option is provided (see the Samples csproj for an example) the issue goes away!